Add gearmand e2e
