+++ gearmand/tox.ini
@@ -5,6 +5,8 @@ envlist =
   py{27,37}
 
 [testenv]
+description=
+    py{27,37}: e2e ready
 dd_check_style = true
 usedevelop = true
 platform = linux|darwin|winn32
